Extract Transformer.execute into a separate file
